basic sound settings panel functionality

+++ GlobalQuakeCore/src/main/java/globalquake/core/Settings.java
@@ -138,7 +138,7 @@ public final class Settings {
 	public static Boolean displayCityIntensities;
 	public static Boolean displayCapitalCities;
 
-	public static Double globalVolume;
+	public static Integer globalVolume;
 
     static {
 		load();
